LayoutTests/fast/dom/HTMLElement/attr-invalid-string.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<div id="result"></div>
<div id="d" contentEditable="abc"></div>
<div id="d2" contentEditable="inherit"></div>
<script>
description('When contentEditable attribute is invalid string, element.contentEditable returns "inherit".')

shouldBe('document.getElementById("d").getAttribute("contentEditable")','"abc"');
shouldBe('document.getElementById("d").contentEditable', '"inherit"');
shouldBe('document.getElementById("d").isContentEditable', 'false');
shouldBe('window.getComputedStyle(d, "").getPropertyValue("-webkit-user-modify")', '"read-only"');

shouldBe('document.getElementById("d2").getAttribute("contentEditable")','"inherit"');
shouldBe('document.getElementById("d2").contentEditable', '"inherit"');
shouldBe('document.getElementById("d2").isContentEditable', 'false');
shouldBe('window.getComputedStyle(d2, "").getPropertyValue("-webkit-user-modify")', '"read-only"');

document.getElementById("d").style.display= 'none';
document.getElementById("d2").style.display= 'none';
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
