third_party/blink/web_tests/platform/generic/fast/dom/HTMLElement/attr-invalid-string-expected.txt

When contentEditable attribute is invalid string, element.contentEditable returns "inherit".

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById("d").getAttribute("contentEditable") is "abc"
PASS document.getElementById("d").contentEditable is "inherit"
PASS document.getElementById("d").isContentEditable is false
PASS window.getComputedStyle(d, "").getPropertyValue("-webkit-user-modify") is "read-only"
PASS document.getElementById("d2").getAttribute("contentEditable") is "inherit"
PASS document.getElementById("d2").contentEditable is "inherit"
PASS document.getElementById("d2").isContentEditable is false
PASS window.getComputedStyle(d2, "").getPropertyValue("-webkit-user-modify") is "read-only"
PASS successfullyParsed is true

TEST COMPLETE
